(PUP-10228) Create NeedLock state for ssl state machine
We want to unlock the lockfile before sleeping, which means we need to reacquire
when we resume. To make that easy, move the locking behavior into a `NeedLock`
state so it can be run at the beginning of the state machine and when resuming.

If locking succeeds, we move to `NeedCACerts` like we did before. If locking
fails, return a `LockFailure` state, which cause a `Puppet::Error` to be raised
the same way it did previously.

We can't use the `Error` state, because it will retry when not running onetime.