(PUP-10390) Ruby file gets loaded twice
In certain environments, the methods and variables in
’puppet/provider/package/gem.rb’ were being initialized twice. First and
only consequences of this issue appeared when constant variables were
added to this file and warnings regarding double
initialisation/definition were shown to the user at every puppet run.

This happened only when puppet’s autoloader from
‘puppet/util/autoload.rb’ parsed and loaded
’puppet/provider/package/puppet_gem.rb’ first and after it did the same
with ’puppet/provider/package/gem.rb’, even though
’puppet/provider/package/puppet_gem.rb’ required
‘puppet/provider/package/gem.rb’ already. The loading order is important
for producing this issue. It can be affected by factors such as
operating system, modules installed and decisional mechanisms used (like
the 'specificity' method from ‘puppet/provider.rb’).

Due to complexity reasons, the final solution was to simply remove the
unnecessary ‘puppet/provider/package/gem’ requirement instead of digging
deeper in the workings of the autoloading mechanism and risking
regressions while trying to improve it. Additionally, because of
similarities, this solution was checked against PUP-9794.