(maint) Move and update config application spec
Move the spec to the application directory.

Rewrite test so it doesn't stub puppet internals, instead pass the tmp
puppet.conf to the initialize_settings method.

Verify we can get/set/delete a key with a mixed UTF8 value. All of our settings
keys are US-ASCII, so remove tests that defined a new settings key as
MIXED_UTF8.