(maint) Update systemd tests
The systemd service integration test only ran on suitable systems, which kind of
defeats the purpose of early feedback. Delete the integration test, restrict the
unit test so it doesn't run on windows. The puppet_service_management beaker
test will be enabled in a different pull request as it hasn't run in awhile.