(PUP-10408) Remove deprecated '-noidme' flag from pkgdmg package provider
The pkgdmg package provider uses the '-noidme' flag with hdiutil, which was deprecated in macOS 10.15. See man hdiutil for more information. In macOS 10.15 the '-noidme' flag causes hdiutil to output a deprecation warning, which then causes pkgdmg to fail to parse the plist xml that hdiutil generates, which then causes the package installation to fail in puppet.

This change removes the '-noidme' flag from pkgdmg in order to prevent the deprecation warning from hdiutil, and restore functionality to pkgdmg in macOS 10.15 and above. Additonally, after reviewing the man page for hdiutil in macOS 10.14, it seems the '-noidme' flag is unnecessary for older versions of macOS as hdiutil appears to not perform idme actions by default. This change also removes the '-noidme' flag from the pkgdmg test in order to bring it up-to-date with the change in pkgdmg.