(maint) Use callback to report on resolver exceptions
Previously, resolver exceptions were collected in the session, and were only
cleared the next time `route_to` was called. To ensure the exceptions have the
same lifetime as the call to `route_to` and no more, pass an error_handler
callback to the `route_to` method. The callback will be called once per
exception that occurs during the resolution process.