(PUP-10385) Ensure filebucket params are included
Prior to this commit, we were not including important parameters in the
new http requests for filebucket. This commit updates the new code to
actually do that.

This also changes some tests. Backing up a file no longer requires the
server to return the Content-Type header. A response body is a bit
strange for a PUT request, but was likely a requirement left over from
the strangeness of the indirector. It's fine that a put request that
doesn't reveive Content-Type back from the server doesn't error.