(maint) Simplify pidlock detection when we already own the lock
If the pidlock file contains our process id, then we by definition own the lock,
and can short-circuit the `clear_if_stale` logic.

It's possible pid was recycled, but that means whatever process created the
pidlock no longer exists, and we've "inherited" the pidlock.

This commit updates `clear_if_stale` to call `lock_pid` only once, as each call
reads the pidlock from disk.

As a result of this change, we only attempt to lookup the process name and args
in the case where the lockfile contains a pid that isn't us and the pid is still
active. As a result, some of the stubbing logic was removed from the spec test.
This also means calling `locked?` in `rake parallel:spec` no longer clears the
lock. A new test wasn't added for that because the "should become locked" test
covers it.