(maint) Fixes comparison of pip versions
When comparing a version with local segment containing letters (eg.
1.2+local01) against the same version but without the local segment (eg.
1.2) an ArgumentError got raised due to an unexpected comparison of
String and Float.

Versions without the local segment got (in the key used for comparison)
a minus infinity float as placeholder. Since the other cases where the
local segment existed (either just as a String, an Integer or any kind
and number of combinations of the two delimited by dot\underscore\dash)
are transformed in pairs of elements (Float and String) in arrays, this
inconsistency caused issues in comparisons such as the case presented
above.