(PUP-10418) Unify ca service return values
This is for Puppet::HTTP::Service::Ca

Unify the return values of Puppet::HTTP::Service methods. Currently we
are not consistent with what these methods return. Some can return a
boolean, some the stringified response body, and some the response
object.

First, all of these methods should return the response so that the
coller can elect to resubmit the request with different parameters if
needed. If the method currently returns an entity, then we should have
it return an array containing `[response, entity]`. This may mean the
response itself is ignored, which is fine.