(PUP-10420) Allow state machine to retry ssl lock
Previously, if the state machine failed to acquired the ssl lock it would raise
a Puppet::Error with the message "Another puppet instance is already running;
exiting" causing puppet to exit.

This commit makes it possible for the state machine to retry the ssl lock up to
a maximum deadline. By default, the state machine won't retry.

If we fail to acquire the lock (after retrying or not), we raise Puppet::Error
and state the reason why we're exiting.