(PUP-10407) Don't chown log file destination
The puppet master process used to drop privileges to the Puppet[:user] and
Puppet[:group], and then reopen its log destinations, so it was necessary that
its log file destination be owned by those accounts.

In puppet 4, the puppet-agent package no longer creates the `puppet` user and
group, only the pouppetserver package does that. As a result, running puppet
agent as root on a host that wasn't the puppetserver would always log a `Unable
to set ownership` error message.

This commit removes the `chown` logic as it's not needed for puppet agents. Also
confine the test based on Puppet::Util::Platform and not features. The latter
requires the features subsystem and is overkill for what's needed here.