(PUP-8051) Address test race condition
This commit modifies the tests for PUP-8051. If we have an integration
test verifying behavior, we shouldn't need the unit test that verifies
the same behavior, so the unit test is removed.

In the integration test, there was a race condition where sometimes for
some slow platforms, the file wouldn't be created and/or written to
within a 0.5 sleep. This commit ensures the file has both been created
and written to before proceeidng, regardless of how long it takes.

Also some formatting cleanup.