(PUP-10416) fix comparision on upgrade/downgrade
When another version of the requested packages is
installed, we do a comparision to determine if we
need to add update/downgrade flag. This comparison
was wrongly done when ranges were provided, conparing
the range string with the version string literally.
This commmit changes the comparasion to be between
the best version in range and the currently installed version.