(PUP-8937) Check AIX package state after install
Previously, when installing packages on AIX systems, Puppet would only
check the package name and version to ensure presence.

If a package is in a broken or inconsistent state, Puppet would not
report any error, causing false positives concerning installed packages.

Extend the package query regular expression to also take into account
the package state, which can be one of the following:

```
A – Applied.
B – Broken.
C – Committed.
E – EFIX Locked.
O – Obsolete. (partially migrated to newer version)
? – Inconsistent State...Run lppchk -v.
```

If the package we are installing gets into a broken (B) or inconsistent
(?) state, we abort the installation. We also treat packages marked with
these codes as absent when querying installed packages.