(PUP-7163) systemd provider 'indirect' services fix
Services that are in the 'indirect' state should not be enabled or
disabled by puppet. The 'systemctl' call with the 'is-enabled' argument
will continue to report the service to be in a 'indirect' state even
after attempting enable/disable it, causing puppet to retry at every run
if the manifest requests so. Before this commit it reported false
changes each time also.

A debug log will be shown each time there is an attempt to enable/disable
an 'indirect' service and no changes will be done.