(PUP-7663) -1 instead if '' when expiry = :absent
When a user is created, useradd will execute 'useradd -e -M test'.
Useradd -e on SLES requires a number of days parameter that is optional on other
platforms. Without the parameter, useradd -e on SLES will create an
expired user with expiration date on 1970-01-01.

The newly added -1 means the account will never expire