(PUP-10292) Evaluate trusted-external lazily
Trusted external command may be expensive to run, so we only want to run
it if we actually want/use the trusted external data. Previously, this
command was invoked on pretty much every call, including file metadata
requests, of which there can be hundreds during some Puppet runs.

After this commit, the trusted external command should typically only be
invoked twice per Puppet run - once when the agent requests a Node, for
environment, to start Pluginsync, and once when the agent requests a
catalog.