(PUP-10479) systemd wrongly reports enabled state
When a systemd service is managed with `puppet resource`,
inconsistencies could be encountered depending on the properties
managed.

`Puppet::Type.to_resource` checks the status of each parameter,
ultimately calling the `enabled?` method in the systemd provider. If
`statuscmd` was previously called, it would have overwritten the
contents of `$CHILD_STATUS`, which is used by the `enabled?` method as
the exit code of `systemctl is-enabled`.

This could happen when passing both `ensure=stopped` and
`enabled=true` to a service, causing `systemctl is-active` to exit
with 3. When `enabled?` is called again, `$CHILD_STATUS` refers to the
context of the `is-active` call, which is not relevant anymore.

To fix this, we update the method calling `is-enabled` to also return
the exit code.