(PUP-10482) Add tests for http file_metadata terminus
Add tests for http file_metadata terminus. Note it returns instances of
`Puppet::FileServing::HttpMetadata` which subclass the terminus' model
`Puppet::FileServing::Metadata`.

Search doesn't make sense for webservers so that method isn't supported.

The code assumes the request URI is already encoded in the manifest, which is
wrong, but we can't change until major version. Mark the test as pending.