(PUP-10482) Use http client to retrieve http(s) file metadata
Include the system store like we do when downloading file content from https
sources.

HttpProxy doesn't explicitly start the connection before calling `head`
on it, which causes ruby to automagically send the `Connection: close`
header, and close the connection when it's done. We preserve that behavior by
setting the header explicitly. The connection pool knows to not cache
the connection when that happens.

The code outputs a slightly different error message, but preserves the
'certificate verify failed' part, so update the integration test.