(PUP-10482) Send Connection: close in test server
The test webrick instance never uses persistent connections, but it did claim
to (by sending `Connection: Keep-Alive`), which could cause problems if a test
tried to reuse a connection.

This commit sets `Connection: close` explicitly, and updates the test that
relied on the server sending `Keep-Alive` but closed the connection.