(PUP-8051) waitforlock and maxwaitforlock setting
Before this commit, an agent run would immediately exit if there was an
already ongoing instance. There was no possibility of making it wait
and retry before doing so. The only workaround was to manually retry
multiple times until the agent lockfile was released.

Two CLI arguments/puppet configuration settings were added in aid of
this issue.

Description for 'waitforlock':
How frequently puppet agent should try running when there is an already
ongoing puppet agent instance. This argument is by default disabled
(value set to 0). In this case puppet agent will immediatly exit if it
cannot run at that moment. When a value other than 0 is set, this can
also be used in combination with the maxwaitforlock argument.

Description for 'maxwaitforlock':
The maximum amount of time the puppet agent should wait for an already
running puppet agent to finish before starting a new one. This is set by
default to 1 minute. A value of unlimited will cause puppet agent to
wait indefinitely.