(PUP-10500) Allow codedir to be a symlink
When using PE with code-manager and `versioned-deploys`, the codedir is a
symlink to a versioned code directory. This commit adds a `symlink_or_directory`
setting to allow the codedir to be either, depending on which mode you're
running in. Without this, puppetserver fails to restart while using
`versioned-deploys` because the settings catalog tries to create a directory
where there is already a symlink.