(maint) Iterable performance improvements
In Puppet's functions involving iteration we call
Iterable.asserted_iterable() to ensure whatever we are given is
properly coerced into an iterator. However, we also do a recursive type
inference on any elements of an Array, or keys and values if passed a
Hash. This is unnecessary for nearly every workflow and can be
prohibitively costly when working with large collections.

The one usage that could use an Iterator with typed elements is if a
user called `step` or `reverse_each` and attempted to save that
value or pass it into a 4x Puppet function defined w/in a .pp file. In
those cases a user _could_ assert that they received an
Iterator[SomeElementType].

To account for this we add a new optional parameter to
`Iterable.asserted_iterable` and `Iterable.on` that allows us to control
whether collection elements are recursively typed. We now only request
elements of a collection to be recursively typed if the user has called
the specific arities of `reverse_each` or `step`.