(maint) Directly delegate known methods in Iterable
Within Iterable we register a method_missing hook to delegate any
methods called on it that have not been defined to the wrapped
enumeration. However, we call several methods within our functions that
we can easily track and know will end up being handled by
method_missing.

Instead of leaving the delegation of methods we need for core Puppet
functions to the method_missing handler this patch directly delegates
for explicitness.