(maint) Inline most common eval paths
The visitors can be relatively slow compared to normal function
invocation. Here we prefer inlining those where it is simple and safe to
do so.

During a catalog compilation most AST nodes have a similar structure and
once in a specific node we know that we will most likely go to one of a
very few different other types of nodes (eg, a VariableExpression almost
always has a QualifiedName for its identifier).

Most catalogs are also dominated by a few different kinds of nodes (eg
LiteralValues, VariableExpressions, ResourceAttributes).

Here is an example from an uncustomized PE Master version 2018.1:
%,    count, eval type
37.9, 33830, "eval_LiteralValue"
18.5, 16505, "eval_VariableExpression"
8.1,  7215,  "eval_AttributeOperation"
5.8,  5195,  "eval_TextExpression"
3.6,  3215,  "eval_ConcatenatedString"
3.4,  3048,  "eval_AssignmentExpression"
2.6,  2316,  "eval_IfExpression"
....

LiteralValues are the most common though they are themselves comprised
of specific types, mostly QualifiedNames and LiteralStrings, eg:

  count base type         -> concrete type
  20394 eval_LiteralValue -> Puppet::Pops::Model::QualifiedName
  12063 eval_LiteralValue -> Puppet::Pops::Model::LiteralString
    932 eval_LiteralValue -> Puppet::Pops::Model::LiteralBoolean
    290 eval_LiteralValue -> Puppet::Pops::Model::LiteralInteger
    182 eval_LiteralValue -> Puppet::Pops::Model::LiteralRegularExpression
      5 eval_LiteralValue -> Puppet::Pops::Model::LiteralFloat

In this patch we've found those places that most frequently lead to
evaluating QualifiedNames, LiteralStrings, VariableExpressions,
AttributeOperations, TextExpressions, and ConcatenatedStrings.

ConcatenatedStrings and TextExpressions were borderline for entry to
this optimization attempt but they also are heavy consumers of
VariableExpressions and LiteralStrings.

In each place we've put a conditional the conditional should catch 80+%
of those objects being tested to offset any slowdown for those that are
not caught. Each place should also contribute significantly to the
number of consumed objects. For example, 80% of QualifiedNames pass
through the VariableExpression evaluation, and 100% of those "name"
field objects within a VariableExpression tested were a QualifiedName.

Classes that would be evaluated as a LiteralValue (QualifiedName &
LiteralString) also simply have their value member called since that is
all their evaluation does.