(maint) Move adapter creation into a constructor
The adapters allow passing a block to `adapt`, however most blocks
contain code that only executes if a value doesn't exist already.
Effectively becoming initialization logic that's executed on every
invocation.

This moves the above kind of logic into an actual constructor for
each concrete Adapter we create. It also moves the remaining portions of
the block passed to `FileNamespaceAdapter.adapt` to simply act on the
returned adapter rather than returning from a nested block.