(PUP-10506) Fix `purge_ssh_keys` requiring homedir
Previously, if `purge_ssh_keys` was set on a user resource similar to:

```puppet
user { 'testuser':
  purge_ssh_keys => true
}
```

Application would fail since the `purge_ssh_keys` parameter requires the
home directory to also be managed by Puppet.

This commit adds a fallback to `Dir.home(username)` for resources that
do not specify the `home` parameter.