(PUP-10289) Allow caller to pass a URL string when they meant a path
The Connection class accepts a URL as the request path, and sends it in
"absolute-form" in the request line[1]:

    GET https://puppet:8140/ HTTP/1.1\r\n

Client's are only supposed to send absolute-form when CONNECT-ing to a proxy so
that the proxy knows which upstream server to forward the request to. However,
HTTP 1.1 servers are required to accept absolute-form, so things happen to work.

This commit allows the ConnectionAdapter to accept a URL like the Connection
class does, and passes a new URL to the client consisting of the current scheme,
server, port and path from the URL argument. The resulting request will be
made using "origin-form" as it should have been all along:

    GET / HTTP/1.1\r\n

There shouldn't be any compatibility issues as servers prefer "origin-form".
This commit also preserves the existing behavior of requiring the caller to URL
encode the string that is passed in, so "/foo%20bar" will work, but "/foo bar"
won't.

[1] https://httpwg.org/specs/rfc7230.html#absolute-form