(PUP-7814) Add ssl_trust_store setting to load trusted CA certs
If `Puppet[:ssl_trust_store]` is not nil and refers to a non-empty file, then
puppet will add that as a trusted CA store when making HTTPS requests to
non-puppet infrastructure, in addition to the puppet CA(s) and CA certs in the
puppet-agent CA cert bundle.

The setting will be ignored if the path refers to an empty file. A warning will
be emitted if the path refers to a directory, etc.