(PUP-1289) Manage service logon credentials on Windows
This feature allows puppet users to gain more control over their Windows
services by now being able to manage the logon credentials for any of
them as desired.

This commit adds the `logonaccount` property and the optional
`logonpassword` parameter (not needed when the user specified in the
`logonaccount` field has no password). These can be managed alone or in
combination with the `ensure` property.

In the first case, the credentials will be checked for validity and then
set but will not be used by the service until it is restarted. This
respects the Windows GUI behaviour from `services.msc` where after
changing the logon credentials, it prompts the user with the alert
message `The new logon name will not take effect until you stop and
restart the service.`.

When also managing `ensure` with it specified as `running` or `true` and
the `logonaccount` is different from the current service settings,
puppet will automatically restart the service to make instant use of the
newly given credentials.

The `logonaccount` field accepts local users (when respecting the
pattern `User`, `.\User` or `PCNAME\User`) and also from domain
(respecting the pattern `DOMAIN\User`, `User@domain.com` or just `User`
if the name doesn't also match with a local one as our user validation
starts looking for its existence and information about it locally first,
and if it's not found there, only then the search will continue onto the
domain). It is not case sensitive and for the problematic language
specific users (such as `System`, `Local Service`, `Network Service`
etc.) their respective SIDs can be used instead (`S-1-5-18`, `S-1-5-19`,
`S-1-5-20` etc.).