(maint) Don't wait when deadline is 0
Time.now.to_i returns the number of seconds since the epoch, so previously it
was possible for `Time.now.to_i` to be equal to `wait_for_lock_deadline` for
values of `maxwaitforlock` between 0 (inclusive) and 1. As a result, the "exits
if maxwaitforlock is exceeded" test always slept 1 second prior to retrying the
lock. However, the background thread only holds the lock for 2 seconds, so on
slow systems it was possible for the background thread to finish and release
the lock before the deadline was reached, causing the test to fail.

To avoid the race condition, if we fail to acquire the lock, compute
Time.now.to_i once, and exit if Time.now is equal to or greater than the
deadline. Apply the same ">=" change to the ssl lock code. As a result of this
change, the flapping test will never attempt to retry the lock.

Alternatively, we could use Time.now.to_f, but ruby has different precision for
different platforms, so let's avoid floating point math.