(maint) Don't wait when deadline is 0
Previously the "exits if maxwaitforlock is exceeded" test sometimes failed
because it started a background thread to lock the lockfile and slept for 2
seconds before exiting. Meanwhile the main thread ran the `agent` application
which can take some time to execute. The 0.5 maxwaitforlock setting also forced
the main thread to always retry the lock once, and that incurred a 1 second
penalty based on `waitforlock`. So if the background thread exited (which
releases the lock) before the main thread retried, then the main thread would
acquire the lock and never output the expected "Exiting now..." message.

This commits sets the maxwaitforlock setting to 0 and changes the inequality to
'>=' to ensure we don't retry the lock when the setting is 0. We also memoize
Time.now, so that we don't cross a second boundary between the time we calculate
the deadline and when we check to see if we reached the deadline.