(PUP-10391) Quiet Ruby 2.7 URI.escape warning
Enable Ruby 2.7 without floating the warning about how URI.escape is obsolete.
After this, we will switch to using the Addressable gem to handle URI
encoding.

With the plan to switch to the Addressable gem to handle URI formatting,
this commit starts to consolidate calls to URI. The idea is we should
have a consistent API, so that when it comes time to switch over to
Addressable, the changes will be minimal.

URI.unesacpe was also deprecated, so also needs to be wrapped so we can
quiet the deprecation until we make the switch to Addressable.

URI.decode is an alias for URI.unescape, so let's sprinkle in some
consistency for fun.