(PUP-10537) Ruby 2.7 warning: keyword param as last arg
`warning: Using the last argument as keyword parameters is deprecated;
maybe ** should be added to the call`

In preparation for Ruby 3, Ruby 2.7 has added warnings around how
keyword and positional arguments are used. Details can be found at:
https://www.ruby-lang.org/en/news/2019/12/12/separation-of-positional-and-keyword-arguments-in-ruby-3-0/

This commit is just covering the changes needed for ssl_provider_spec.rb, though it might be a better idea to merge these commits.