(maint) Don't retry the lock if the deadline is now
Commit c5a8687ffee intended to switch the comparison to >= so that if the
deadline is the same as now (in integer seconds since the epoch) then we
shouldn't retry the lock. If the values were equal then the test would sometimes
fail on slow systems (windows appveyor) because the background thread exited and
released the lock before the retry occurred, so it didn't exit with the expected
message.