(PUP-10511) fix resource array with sensitive parameters
In case of resource arrays, the parameters are shared between resources

```
test_mod::service { ["ServiceA","ServiceB"]:
  password => Sensitive("password")
}
```

This leads to trouble for Sensitive parameters as puppet will unwrap the
value for first resource and will validate it against Sensitive type for
the second resource.

With this change, the parameter and it's value are duplicated when
resource is created.