(PUP-10527) Test puppet service runs puppet
This commit adds an acceptance test to ensure that starting the puppet
service on agents runs puppet.  This validation was missing from the
`puppet_service_management.rb` test.  Due to the comprehensive nature of
the existing test, this test is added independently to reduce execution
time.  The existing test takes about 4m to execute while this specific
case takes about 30s.