Add debug information when SSL peers are being verified
It's useful when debugging the SSL configuration of agents to have an
easy way to debug what's the SSL peer being verified and the trusted CA
certificates being used.

This patch adds a debug message indicating the subject of the peer being
verified, the subject of each CA certificate being used to verify
signatures with and the result of the vefification for each pair.