(PUP-10254) Add wide terminator when converting to pointer to wide string
When copying a wide character string from the ruby heap to the native heap, we
allocated a buffer equal to the number of bytes of the wide string, but did not
append a wide terminator. We were getting lucky as the prior call to
`wide_string` embeds a wide terminator in the string data.

Now allocate a buffer to include the wide character terminator, and ensure it is
terminated. This matches `FFI::MemoryPointer.from_string` behavior for single
byte characters, which allocates n+1 bytes[1] and null terminates the string[2].

[1] https://github.com/ffi/ffi/blob/b54e735b11abef310c58b0f33dfc1af8f78f007e/ext/ffi_c/MemoryPointer.c#L163-L165
[2] https://github.com/ffi/ffi/blob/b9248848ad52893ffe1ad0ac6e97f207b0119346/ext/ffi_c/AbstractMemory.c#L519