(PUP-6380) Reset Host header when redirecting
Previously we were preserving the Host header across redirects, which is
incorrect if we've been redirected to a different host. Ruby will automatically
update the Host header, including omitting the default 443/80 port, but it's
better to be explicit.

Note Amazon presigned URIs will fail if the default port is included in the Host
header after a redirect.