(PUP-6380) Apply query parameters from the Location header
Previously, we used the query parameter from the base request URI instead of
the Location header. Also we did not follow relative redirects correctly. Use
Ruby's URI.join method[1] to resolve the Location URI relative to the base
request URI. If the Location URI is absolute, then we'll just use that.
Otherwise resolve the new path, query and fragment as specified in RFC2396.

[1] https://github.com/ruby/ruby/blob/v2_5_7/lib/uri/generic.rb#L1100