(PUP-5707) Support X-Checksum-* headers
Artifactory generates X-Checksum-{Md5,Sha1,Sha256} headers, so parse them from
the HTTP response to determine the remote checksum. Note we parse all of the
checksum headers, but use the first header based on the order given in the
`collect` method.

This commit is taken from PR https://github.com/puppetlabs/puppet/pull/5707.

There is a draft RFC for Want-Digest and Digest headers, but it's not finalized
yet: https://tools.ietf.org/html/draft-ietf-httpbis-digest-headers. At some
point, we'll probably want to support that.

Co-Author: Dylan Cochran <heliocentric@gmail.com>