(PUP-5707) Add sha1 and sha1lite as valid checksum types
Previously, the sha1 and sha1lite checksum types were not included when defining
the list of valid checksum types for the file resource. However, those checksum
types could still be used, because the `Puppet::Type::File` uses
`send("sha1?", value)` to check if a checksum is valid[1], which ends up calling
the `sha1?(value)` private module method from the `Puppet::Util::Checksums`
module[2] that is mixed into `Puppet::Type::File`[3].

To ensure we have a single source of truth, splat the known checksum types when
calling `newvalues`. The order of values passed to `newvalues` is different, but
it is not significant. For example, the http filemetadata class collects
checksums in a particular order, and only those that correspond to an HTTP
header.

Also make the known checksum types a constant and remove extraneous parentheses.

[1] https://github.com/puppetlabs/puppet/blob/6.16.0/lib/puppet/type/file.rb#L431
[2] https://github.com/puppetlabs/puppet/blob/6.16.0/lib/puppet/util/checksums.rb#L8
[3] https://github.com/puppetlabs/puppet/blob/6.16.0/lib/puppet/type/file.rb#L16