(PUP-5707) Add valid_checksum? method
Previously, `Puppet::Type::File` used `send` to validate checksum types and
values, calling a private module method on the `Puppet::Util::Checksums` module
that is mixed into `Puppet::Type::File`. This is a bit fragile because the
caller knows how to construct the name of the predicate method, eg "sha1?",
which makes refactoring more error prone.

This commit adds a `valid_checksum?` method to make it explicit. It preserves
the behavior of raising a NoMethodError if the checksum type is unknown.