(PUP-10544) Request protocol should take precedence over path
Previously, if an indirector request's key was a URL, but the URL path started
with a double slash, then puppet thought it was a local absolute path. This
occurred because the indirector request strips the leading slash off the URL
path when setting its key, but the second slash remained, and that appeared to
be an absolute posix path.

This commit eliminates the premature optimization. Select the terminus based on
the request protocol. If no protocol is given and the request key is an absolute
path, then use the local :file terminus. Otherwise select the :file_server
terminus, which knows how to convert a mount path to a path on disk, such as
"modules/my_module/my_file".