(PUP-3970) Remove wide string terminator workaround
Ruby < 2.1 did not correctly handle multibyte terminators when allocating and
resizing wide strings, causing data to be read past the end of the buffer. It
was fixed in ruby 2.1.0[1], which we no longer support.

Ruby < 2.4 did not correct handle multibyte terminators when transcoding
strings, such as when calling `String#force_encoding('UTF-16')` and converting
the binary string with a single null terminator to a wide string with a wide
terminator. That was fixed in 2.4.0[2] and backported to 2.3 in [3], which is
the minimum version we support.

[1] https://github.com/ruby/ruby/commit/1549894a0699abdfab1fbba0e468afd3f2f990b3
[2] https://github.com/ruby/ruby/commit/f2ee22371b15411e818b23199161d1658ba08622
[3] https://github.com/ruby/ruby/commit/6f122a4f10512bb60795faa0021e498800a7f1b8