(PUP-10254) Max key and value name lengths should be chars not bytes
The reg_enum_key and reg_enum_value functions incorrectly defaulted the max name
length to bytes (514 & 32768) instead of characters. Per the RegEnumKeyW[1] and
RegEnumValueW[2] docs, the `lpcchName` and `lpccValueName` parameters expect to
receive "the size of the buffer specified ... in characters... including the
NULL terminator". This wasn't an issue in practice because the caller always
calls `RegQueryInfoKeyW` to compute the max character lengths for that hive. But
to avoid future confusion, change the default max to be buffer lengths in
characters based on [3].

Note ruby's `Win32::Registry::MAX_KEY_LENGTH` is 514 = (255 + 1 + 1) * 2. It's
not clear why ruby is including two wide NULL characters. Likely they are
confusing that with a REG_MULTI_SZ data containing a double wide NULL.

[1] https://docs.microsoft.com/en-us/windows/win32/api/winreg/nf-winreg-regenumkeyexw
[2] https://docs.microsoft.com/en-us/windows/win32/api/winreg/nf-winreg-regenumvaluew
[3] https://docs.microsoft.com/en-us/windows/win32/sysinfo/registry-element-size-limits