(PUP-10254) Detect null terminators at the end of the array
Previously, we used an exclusive range to scan wchar indexes from
0 to `max_char_length - terminator_width`. However, due to an off-by-one
error, we never checked the last wchar in the array.

For example, if the array contained wchars ['A', 0] and max_char_length
was 2, then we computed the range `(0...2-1) = (0...1)`, which
corresponds to the array `[0]`. So the wchar at index 1 was skipped.

With this commit, the range is computed as `(0...2)` which corresponds
to the array `[0, 1]`.

The old code happened to work, because we called `read_wide_string(2)`,
which creates a 2 wchar string containing `A` and an embedded wide null.
However, transcoding from UTF-16LE to UTF-8 strips the trailing nulls.

When scanning for a double-wide null in the wide string "foo\0\0", we
decrease the upper bound by one, so that we don't read out of bounds
trying to read 32 bytes. For example, given `max_char_length` of 5, we
scan wide character indexes 0, 1, 2, 3, and we correctly detect the
32-byte 0 at index 3.