(maint) Add Windows predefined local account check methods
This commit adds the ability to check if a user is one of the predefined
local accounts on Windows. The checks can be done for: 'LocalSystem', 'System',
'Local Service' and 'Network Service'. These accounts don't have a password
and any means of checking their password through logon attempts will
fail.

'LocalSystem' cannot be specified in a call to the `LookupAccountName`
function, but since the name is the same in all locales (only in GUI
it's spelled differently based on the OS language), there is no need to
compare it with a well known SID when checking it.